Add a prototype of a quadtree
I need something to check where the player can walk and where not. First
idea was to create for each tile of a level a 'bounding box' and check
the players bounding box against them. The level 'bounding boxes' should
be stored in a quadtree, because you don't want to check a Set...
After a while of thinking, I realized at the moment a Grid would solve
the same problem. (so I will keep the prototype, perhaps for later).