Add basic functionality for collision detection between an entity and the level
This commit, did also some other stuff. I cleaned up the example resource file structure
to avoid duplicate files. I cleaned up the whole Entity (the basic Entity Component Model).
It more like a state model which can be customized (Builder Pattern) with certain 'states'.
The drawable is the exception, because it includes stuff for rendering. I decided to have
this 'state model' because I struggled with -where to put the functionality- like movement,
physics, etc. At the moment it is good enough to work with. I also created Example05, which
should demonstrate the collision detection (with jump mechanics), but at the moment it's not
finished. I also created the GridGenerator tool for creating a Grid (walkable paths, solid
blocks) which can be generated from the level block file (where the level tiles are generated
from -> LevelTilesGenerator)