Change the collision detection for the player
Feels a bit better, but it's not perfect at all. Very rarely the player
get stuck inside the ground (I don't know why ^^). Now I don't use the
next tick (to look into the future, where the player position would be),
instead I let the player collide and snap him afterwards to the Grid.
Well... I'm not really satisfied with it, but it works a bit better
than before.